// src/application/dtos/subscription/subscription-dto.ts
export interface CurrentUsageStatsDTO {
  profileId: string;
  shopId?: string;
  todayQueues: number;
  currentShops: number;
  currentStaff: number;
  monthlySmsSent: number;
  activePromotions: number;
}